+++ src/components/ProductListItem/index.tsx
@@ -6,49 +6,16 @@ import * as React from "react";
 import { Thumbnail } from "@components/molecules";
 
 import { TaxedMoney } from "../../@next/components/containers";
-import { BasicProductFields } from "../../views/Product/gqlTypes/BasicProductFields";
-
-export interface Product extends BasicProductFields {
-  category?: {
-    id: string;
-    name: string;
-  };
-  pricing: {
-    priceRange: {
-      start: {
-        gross: {
-          amount: number;
-          currency: string;
-        };
-        net: {
-          amount: number;
-          currency: string;
-        };
-      };
-    };
-    priceRangeUndiscounted: {
-      start: {
-        gross: {
-          amount: number;
-          currency: string;
-        };
-        net: {
-          amount: number;
-          currency: string;
-        };
-      };
-    };
-  };
-}
+import { FeaturedProducts_shop_homepageCollection_products_edges_node } from "../ProductsFeatured/gqlTypes/FeaturedProducts";
 
 interface ProductListItemProps {
-  product: Product;
+  product: FeaturedProducts_shop_homepageCollection_products_edges_node;
 }
 
 const ProductListItem: React.FC<ProductListItemProps> = ({ product }) => {
   const { category } = product;
-  const price = product.pricing.priceRange.start;
-  const priceUndiscounted = product.pricing.priceRangeUndiscounted.start;
+  const price = product.pricing?.priceRange?.start;
+  const priceUndiscounted = product.pricing?.priceRangeUndiscounted?.start;
 
   const getProductPrice = () => {
     if (isEqual(price, priceUndiscounted)) {
@@ -70,7 +37,7 @@ const ProductListItem: React.FC<ProductListItemProps> = ({ product }) => {
         <Thumbnail source={product} />
       </div>
       <h4 className="product-list-item__title">{product.name}</h4>
-      <p className="product-list-item__category">{category.name}</p>
+      <p className="product-list-item__category">{category?.name}</p>
       <p className="product-list-item__price">{getProductPrice()}</p>
     </div>
   );